ai/pyproject.toml
[tool.poetry]
name = "gayradarai"
version = "0.1.0"
description = "GayRadar AI"
authors = ["AveryanAlex <alex@averyan.ru>"]

[tool.poetry.dependencies]
python = "^3.11"
insightface = "^0.7.3"
pillow = "^10.2.0"
scikit-learn = "^1.4.0"
opencv-python = "^4.9.0.80"
numpy = "^1.26.3"
onnxruntime = "^1.16.3"
fastapi = "^0.109.0"
uvicorn = "^0.27.0"
python-multipart = "^0.0.6"

[tool.poetry.group.dev.dependencies]
black = "^23.12.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
